app/requirements.txt
aniso8601==1.1.0
Flask==0.10.1
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.2
pytz==2016.3
six==1.10.0
Werkzeug==0.11.5
Flask-SQLAlchemy==2.1
passlib==1.6.5
PyJWT==1.4.0
